+++ include/general.hpp
@@ -1,12 +1,12 @@
 #ifndef GENERAL_H
 #define GENERAL_H
 
-double xlinearInter(
-  const double x1,
-  const double y1,
-  const double x2,
-  const double y2,
-  const double y
+float xlinearInter(
+  const float x1,
+  const float y1,
+  const float x2,
+  const float y2,
+  const float y
 );
 
 #endif //GENERAL_H
